+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- new config `lang` to allow setting the number format to be used when printing
+
 ## [v0.52.0] - 2024-06-02
 
 ### Added